fix windows support
- fix broken windows support

+++ backend/amd/minethd.h
@@ -30,7 +30,7 @@ private:
 	// Bottom 24 bits allow for an hour of work at 4000 H/s
 	inline uint32_t calc_start_nonce(uint32_t resume)
 	{
-		return reverseBits<uint32_t>(iThreadNo + GlobalStates::iThreadCount * resume);
+		return reverseBits<uint32_t>(static_cast<uint32_t>(iThreadNo + GlobalStates::iThreadCount * resume));
 	}
 	
 	void work_main();
